db changes

+++ src/com/webservices/services/DailyBasisService.java
@@ -10,5 +10,5 @@ public interface DailyBasisService {
 	
 	public DailyService storeDailyBasisService(DailyService dailyService);
 	public List<DailyService> getDailyBasisService() ;
-	public  List<DailyService> findByDatesBetween(Date startDate,Date endDate,Long serviceId);
+	public  List<DailyService> findByDatesBetween(Date startDate,Date endDate,Integer serviceId);
 }
